Telegram Bot/Omdb/requirements.txt
pip==24.0
python-telegram-bot==20.3
